+++ package.json
@@ -12,6 +12,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@material-ui/core": "^4.9.14",
+    "@material-ui/icons": "^4.9.1",
+    "@material-ui/lab": "^4.0.0-alpha.53",
     "axios": "^0.19.2",
     "babel-jest": "^26.0.1",
     "core-js": "^3.6.5",
